T8 speed boost (includes a new installer (0.0.3))
Change the bootloader's configuration of the MC68377's Periodic
Interrupt Timer to interrupt more often which in turn makes it send CAN
frames more frequently.

No need to read or write all of those 0xFFs at the end of the FLASH
chip. Use 'header' pointer stored at address 0x020140 to find the area
of the FLASH chip that is actually used, adding 0x200 bytes to be sure
of including the header itself.

Skip 'Requesting mandatory data' since it's not a mandatory step! Tech 2
requests this information to check that the update file matches the car
data - but we don't check any of this and sometimes want to use
different BIN files :-0