Legion; Changed how main is flashed++
It's now using a mix of selected partitions and end address to determine
last address while flashing Main.

Added 3 new regions for md5 of main:
10: 0 to last address + 512 bytes
11: 0x4000 to last + 512
12: 0x20000 to last + 512.

Started on a function that, eventually, should show better information
as to why a frame was re-sent during flashing. (And abort if the fault
is of critical nature)

Massive cleanup of bootloader