LegionT8; MCP: Changed behaviour of formatmask
If you start a session that writes boot and it for some reason fails
halfway through a dangerous situation could previosusly occur where
shadow would not be selected again when you retry (hence being left
filled with 0xFF).

This change will make sure both are selected if any of them are
different.