T8leg: Tweaked behaviour of verification
T8 main has a feature where it'll only write to the last used address.
That in itself is neat but it has a tendency to make the regular
per-partition verification grumpy since quite a few bin files has
"hidden" data after the header region that is not used.

That lead to the inception of "CompareRegmd5"
Which also had a few problems of its own:
If only nvdm 2 is tagged for format it will be discarded since nvdm 1 is
considered the master copy.
As a result region 10 and 11 would fail verification since nvdm is part
of those  ranges.

In the end it was easier to just check those individually and only hash
region-md over app data since it has no strange dependecies, user
options or other unknown factors.